Change AABB class to BoundingBox
- Change AABB to BoundingBox
- change OctreeNode box to new BoundingBox
- Remove most warning related to types conversions

+++ ptLeftarCore/Graphics/Color.h
@@ -86,11 +86,11 @@ public:
 static inline Color blendColor(Color a, Color b, float blendingFactor) {
   Color outColor{0, 0, 0};
   outColor.r =
-      sqrt((1.0f - blendingFactor) * a.r * a.r + blendingFactor * b.r * b.r);
+      sqrtf((1.0f - blendingFactor) * a.r * a.r + blendingFactor * b.r * b.r);
   outColor.g =
-      sqrt((1.0f - blendingFactor) * a.g * a.g + blendingFactor * b.g * b.g);
+      sqrtf((1.0f - blendingFactor) * a.g * a.g + blendingFactor * b.g * b.g);
   outColor.b =
-      sqrt((1.0f - blendingFactor) * a.b * a.b + blendingFactor * b.b * b.b);
+      sqrtf((1.0f - blendingFactor) * a.b * a.b + blendingFactor * b.b * b.b);
 
   return outColor;
 }
